requirements.txt
tensorflow==2.18.0
fastapi[standard]==0.115.7
uvicorn==0.34.0
matplotlib==3.10.0
numpy==2.0.2
pillow==11.1.0
keras==3.8.0